pyproject.toml
[project]
name = "agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph-cli[inmem]>=0.4.12",
    "rich>=14.3.1",
]
